Closes #41 : groups stuff
Added basic user's group management.

A user can create a new group if he not a member of any group.
Every user has access to the group info and only a group owner has access to group management.
When a user creates a new group he becomes a group owner and group member.

Groups::Creator is used to create a new group and promotes a user as the group owner.
Api::V1::GroupsController is used to access/control user's group information.