Closes #91 : Add authorization system
Action Policy gem is used to perform all authorization scenarios in application

 - All policies located in the `app/policies`;
 - Policies naming should be with using the corresponding singular resource name (model name) with a Policy suffix,
   e.g. Course -> CoursePolicy;
 - Policies rules naming should be with using a predicate form of the corresponding activity (typically, a controller's action),
   e.g. `CourseController#update` -> `CoursePolicy#update?`.

Rules must be public methods on the class. Using private methods as rules will raise an error.

A Policy is instantiated with the target `record` (authorization object) and the authorization context (by default equals to `user`)

When authorization is successful (i.e., the corresponding rule returns `true`), nothing happens,
but in case of an authorization failure `ActionPolicy::Unauthorized` error is raised