Fix CVE-2019-16109
Devise vulnerability

An issue was discovered in Plataformatec Devise before 4.7.1.
It confirms accounts upon receiving a request with a blank confirmation_token,
if a database record has a blank value in the confirmation_token column.

(However, there is no scenario within Devise itself in which such database records would exist.)