Closes #245 Add audit events
An audit event is any security-relevant occurrence in the system.
A security-relevant occurrence can be a change to the security state of
the user's account.

The information that is reported in an audit event includes the type of the event,
and any additional event-specific information that is related to security auditing.